Update overrides documenation
It seems like the statement:

"Note that the overrides field can only be set at the root of the project."

May, not be a true statement.

After discovering a version conflict in our deployments, we applied the overrides to two packages within our monorepo. This change resolved the version conflict.

I'm sure this statement is here with reason. Do we know why this line is here and why overrides only apply (perhaps a better choice of words is "should") in the root package.json?