Update docs/installation.md (#331)
* Update installation.md 

My contribution with the steps to create an alias for the pnpm command in Powershell in Windows.

This updates the "Using a shorter alias" section.

Note: I think this is my first contribution on Github, I'm excited to be part of this. If any issue comes up, let me know to research into it and help solve it.
I carefully confirmed this solution in the official Powershell documentation. My alias has been working in Windows Terminal Powershell terminal, and in VSCode Powershell terminal, and after system shutdown, and in multiple windows and tabs.

Co-authored-by: Zoltan Kochan <z@kochan.io>